+++ tbgpp-execution-engine/include/execution/physical_operator/physical_produce_results.hpp
@@ -11,9 +11,9 @@ class PhysicalProduceResults: public CypherPhysicalOperator {
 public:
 	PhysicalProduceResults(Schema& sch)
 		: CypherPhysicalOperator(PhysicalOperatorType::PRODUCE_RESULTS, sch) { }
-	PhysicalProduceResults(Schema& sch, vector<uint8_t> projection_mapping)
+	PhysicalProduceResults(Schema& sch, vector<uint64_t> projection_mapping)
 		: CypherPhysicalOperator(PhysicalOperatorType::PRODUCE_RESULTS, sch), projection_mapping(projection_mapping) { }
-	PhysicalProduceResults(Schema& sch, vector<vector<uint8_t>> projection_mappings)
+	PhysicalProduceResults(Schema& sch, vector<vector<uint64_t>> projection_mappings)
 		: CypherPhysicalOperator(PhysicalOperatorType::PRODUCE_RESULTS, sch), projection_mappings(projection_mappings) { }
 	~PhysicalProduceResults() { }
 
@@ -29,8 +29,8 @@ public:
 	std::string ToString() const override;
 
 private:
-	vector<uint8_t> projection_mapping;
-	vector<vector<uint8_t>> projection_mappings;
+	vector<uint64_t> projection_mapping;
+	vector<vector<uint64_t>> projection_mappings;
 	mutable uint64_t num_nulls = 0;
 };
